Talkie (Version 1.0.1
This is basically the first release of the Talkie app! This app allows you to communicate with friends and strangers in a local vicinity without WiFi or Cellular Data! It supports text up to 400 characters, images, user name / profile pic, background image, and is even encrypted! The best part is that there are no ads, no data collection on the app, and all features in the app are free! Just please support us by paying it forward on GameJolt... of all places 😂 https://gamejolt.com/games/Talkie/1022714