Create README.md
This is basically the first release of the Talkie app! This app allows you to communicate with friends and strangers in a local vicinity without WiFi or Cellular Data!
We made Talkie because we wanted a way for people to communicate privately and locally without wifi or cell service. In the past, governments have disabled these routes of cominication or removed the platforms that host communications.
We want Talkie to be the platform you can't easily take down. Where people can share info without being tracked, scanned, or stolen. Thats why we plan to take Talkie to more than just MacOS, we plan to port it to iOS, Android, Windows, and Linux.
We hope you support us in this journey to making the world a more free place.